refactor: Refactor project structure and file naming conventions
- Update project metadata and dependencies in `pyproject.toml`
- Move and rename files to new directory structure
- Add new files and classes to improve code organization
- Update import statements and fix issues in various files
