+++ esg_bot/__main__.py
@@ -1,0 +1,3 @@
+from . import run_chatbot
+
+run_chatbot()
