+++ esg_bot/data_handler/lc_docstore_handler/docstore_handler.py
@@ -1,5 +1,6 @@
 from abc import ABC, abstractmethod
 
+
 class DocstoreHandler(ABC):
     def getStore(self):
         return self.docstore
