+++ esg_bot/data_handler/lc_vectorstore_handler/pinecone_vectorstore_handler.py
