+++ esg_bot/data_handler/lc_vectorstore_handler/vectorstore_handler.py
