+++ esg_bot/data_handler/report_handler/pdf_image_report_handler.py
