+++ esg_bot/data_handler/utils/image_base64_utils.py
