+++ pyproject.toml
@@ -1,0 +1,76 @@
+[tool.poetry]
+name = "esg_bot"
+version = "0.1.0"
+description = ""
+authors = [
+  "Park Jong Hyun <jonhpark7966@gmail.com>",
+  "Jee Wan Gue <jeewangue@gmail.com>",
+]
+license = "UNLICENSED"
+readme = "README.md"
+
+homepage = "https://github.com/jeewangue/esg-bot"
+repository = "https://github.com/jeewangue/esg-bot"
+
+packages = [
+  { include = "esg_bot", from = "." },
+]
+
+[tool.poetry.dependencies]
+python = ">=3.11, <3.12"
+python-dotenv = "^1.0.1"
+pandas = "^2.2.2"
+langchain = "^0.2.11"
+langchain-community = "^0.2.10"
+langchain-experimental = "^0.0.63"
+langchain-openai = "^0.1.19"
+langchain-pinecone = "^0.1.3"
+langsmith = "^0.1.93"
+streamlit = "^1.37.0"
+streamlit-feedback = "^0.1.3"
+finance-datareader = "^0.9.93"
+plotly = "^5.23.0"
+
+[tool.poetry.group.dev.dependencies]
+jupyterlab = "^4.2.4"
+black = { extras = ["d"], version = "^23.1.0" }
+isort = "^5.10.1"
+line-profiler = "^4.0.2"
+memory-profiler = "^0.61.0"
+rope = "^1.7.0"
+mypy = "^1.2.0"
+types-pyyaml = "^6.0.12.12"
+ruff = "^0.5.5"
+pipx = "^1.6.0"
+
+[tool.ruff]
+line-length = 160
+target-version = 'py311'
+# Enable Pyflakes `E` and `F` codes by default.
+select = ["E", "F", "W", "N", "B", "I", "NPY", "PL", "C90", "UP", "SIM", "RUF"]
+
+[tool.ruff.mccabe]
+# Unlike Flake8, default to a complexity level of 10.
+max-complexity = 12
+
+[tool.black]
+line-length = 160
+target-version = ['py310', 'py311']
+
+[tool.mypy]
+# be strict
+disallow_untyped_calls = true
+warn_return_any = true
+strict_optional = true
+warn_no_return = true
+warn_redundant_casts = true
+warn_unused_ignores = true
+
+# The following are off by default.  Flip them on if you feel
+# adventurous.
+disallow_untyped_defs = true
+check_untyped_defs = true
+
+[build-system]
+requires = ["poetry_core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
